+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - `br-nfse-v1`: added initial Brazil NFS-e addon
+- `bill`: normalize line discounts and charges to remove empty rows.
 
 ### Changed